SQL Bug fixes, and additions
- Repaired CSO lookup

- Refactored
-- Refactored SQL to make it more modular
-- Modularised in the hope the Query Optimiser will improve things

Data-sets of uploaded files are pushed out a new view and accompanying Function - the function is there to lighten the load when retrieving a specific submission and the new view is there to extract the code for this to a new view in the hope of query optimisation

The Fetch SP has had it's filtering logic optimised, to reduce the number of steps to achieve an outcome.

The OrganisationRegistration Summaries view has been modularised and reduced in size and retreives CSO infomation accurately.

The OrgansiationRegistrationDetails SP has been heavily refactored.
Firstly, it now uses the Compliance Scheme information properly.
Secondly, it no longer relies on the View to obtain it's core submission information.
It uses the new Function to retrieve upload file details for the download links.