Added a Function that is missing from the repo, with the repair required. (#149)
Also, a view is updated in the repo that contains the same issue.
Both should be deployed to repair PRE PROD.
It has been seen to be a problem to deploy the function and that was previously deployed manually.

Both scripts can be run directly - they will drop and create each object.