[601604] Adjust Subsidiary counts to align with new rules - (#248)
Which are to ignore the leaver_date for subsidiaries when counting them in submission and resubmission scenarios.
The primary change is to remove the filter '-- and leaver_date is null' on the filter for subsidiaries.

This commit ALSO adjusts the fetch org reg details SP to align it with what is currently in production - the performance enhancements made after R13 made it to main but not the R13 branch.